Fix patchops to add rather than replace
The vanilla PDW and Piercer do not have tools in their defs, so trying to patchopreplace them fails, which kills the sequence. This is a quick fix that doesn't rely on changing any of the base defs. If you've already patched the base that these guns may inherit from, just remove the two replace ops instead.