English proof readings (#72)
* Update Sergeant.xml (#69)

Not sure about the Falcon, please correct me if I'm wrong.

* Update Ronin.xml (#68)

Another touchup.

* Update Protector.xml (#67)

Cleaned it up.

* Update Misc.xml (#65)

Quickest work ever. Capitalisation as usual.

* Update MilitaryGun_Mount.xml (#64)

Most of this was capitalization once a gain, with some rephrases added here and there.

* Update MilitaryGun_Light.xml (#63)

Did some touchups here and there, and some phrasing was changed.

* Update MilitaryGun_Heavy.xml (#62)

Did the same thing, just changed some sentence here and there to not be repetitive. Most of the work is capitalization.

* Update MilitaryGun_Gatling.xml (#61)

Did the same changes here, similar to MilitaryGun.xml

* Update MilitaryGun.xml (#60)

I did the standard capitalization of everything, and changed specially into specifically as that might be the word you are looking fore, and replaced "normal" with the phrase "standard baseline" to make it sound official and cutting.

* Update Military.xml (#59)

Proofread. Had to create variations of the descriptions for the soft caps to not sound repetitive and robotic.

* Update DMS_Security.xml (#58)

I just only worked on the present descriptions and did not touch the empty descriptions for the other turrets. Just tell me if you want me to write one or proofread new entries.

* Update Races_Automatroid_Medium_Arquebusiers.xml (#30)

Capitalization and restructuring of sentence, especially for the Mech Sniper description.

* Update Weapon_Melee.xml (#33)

* Update DMS_Production.xml (#57)

Proofread in English.

* Update DMS_Misc.xml (#56)

Proofread in English

* Update DMS_Mechlink.xml (#55)

Proofread in English.

* Update DMS_MechBand.xml (#54)

Proofread with English.

* Update DMS_Manufactured.xml (#53)

Proofread in English.

* Update DMS_Machine.xml (#52)

Proofread in English.

* Update DMS_Drug.xml (#51)

Proofread in English.

* Update CivilianGun.xml (#50)

Proofread in English.

* Update Civilian.xml (#49)

Proofread in English.

* Update DMS_Thoughts_Titles.xml (#48)

Proofread in English

* Update TraderKinds.xml (#47)

Proofread in English.

* Update Shuttle.xml (#46)

Proofread in English.

* Update DMS_Bionic.xml (#45)

Proofread in English.

* Update DMS_Boss.xml (#44)

Proofread in English

* Update Common.xml (#43)

Some proofreading.

* Update KeyedFile.xml (#42)

Capitalisation and grammar correction.

* Update Races_Automatroid_Heavy_Miscs.xml (#41)

Another quick work.

* Update Races_Automatroid_Heavy_Iguana.xml (#40)

Again, some rephrasing, and I added an additional blurb for the Iguana to make sense of its mass production costs.

* Update Races_Automatroid_Heavy_BattleFrames.xml (#39)

Okay, this one was were I had to do a considerable rewriting of the descriptions so that it flows naturally in the eyes during reading and more importantly as an editor, sounds natural to the ears when read aloud.

* Update Races_Automatroid_Heavy_Geochelone.xml (#38)

This one was much simpler, though I had to capitalize here and there and change a word or two.

* Update Races_Automatroid_Heavy_Tarbosaurus.xml (#37)

Okay, this one had to really change the description as it was clunky and awkward grammatically. 

Now it flows better in reading and to the ears when spoken aloud.

* Update Races_Automatroid_Medium_Ronins.xml (#36)

Okay, please correct me, but since the Ronin and Phalanx had the same description except for the Phalanx to be cheaper, I added a small blurb that it is the cheaper variant of the Ronin just to be sure and remain lore consistent.

* Update Races_Automatroid_Machines.xml (#31)

Another quick work, though I had to double check if I double-dipped with the Automatroid_Light.xml file.

* Update Shuttle.xml (#32)

* Update Races_Automatroid_Medium_Soldats.xml (#34)

This basically feels like the descriptions are duplicated. I would like to ask how the Grenadier and Kanonier are different so I can inject some unique blurb that sets them apart.

Either way I cleaned it up and made it more sensible in my opinion.

* Update Races_Automatroid_Light.xml (#35)

Another quick work, had to do some sentence restructuring.

---------

Co-authored-by: bryanfran42 <bryanfran42@gmail.com>